pyproject.toml
[tool.poetry]
name = "verify_domain_cert"
version = "0.1.3"
description = "Verify if a domain certificate has expired"
authors = ["JenGuo Zhai <zhaifengguo@gmail.com>"]
readme = "README.md"
homepage = "https://github.com/zhaifg/check_ssl"
repository = "https://github.com/zhaifg/check_ssl"
packages = [{ include = "verify_domain_cert" }]

[tool.poetry.dependencies]
python = "^3.11"
pyopenssl = "^24.0.0"
idna = "^3.6"


[tool.poetry.scripts]
verify_domain_cert = "verify_domain_cert.check:main"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
